requirements.txt
python-telegram-bot==21.5
anthropic
supabase
python-dotenv
apscheduler
stripe
fastapi
uvicorn